Defer config loading until command is called
We did it this way to prevent constantly reloading the config if the
program remained open (a use case we don't yet support), however it
meant that tests couldn't override config behavior.

Now we use a global singleton to prevent constant reloading but also
defer the loading to a point that tests can affect the behavior.