Add hooks to detect if a relog has occurred
Now that HTTPClient.refresh_token can perform login attempt which will
give a new set of a tokens we needed to provide a way to listen for this
event. refresh_hook is called with the new access and refresh token if
the login was successful.

We now propagate this up from the API class so the config can reflect
those changes if this occurs. Additionally API will use it's own hook to
update it's clients with the new tokens.