feature: support context switching
This adds application state and application configuration to the
Taskforge CLI. As of now the configuration is underutilized (read: not
used at all).

Application state represents things that are not part of a user's
configuration as they are likely to change often and be changed by the
application itself.

Using Application state we now are able to leverage context switching so
that when a user runs task next they get a task that's most relevant to
their current situation. This feature was just added to the backend and
the SDK.