feature: Add the edit command
To make the editing user-friendly we populate the tasks before dumping
to yaml. We then do a reverse-populate to prep them for updating.

To support the reverse-populating the ModelClient added support for a
reverse_lookup that lets you get an object previously retrieved by ID
via some other field.

This makes owner basically read-only (as it is in the API anyway) and we
need to surface this to the user better but it works for now.